Function Docs

This list is generated automatically, which may contain some internal unexported functions.

PureLibm.CR_ATANF_CD — Constant

Denominator coeffs for rminimax.

See CR_ATANF_CN

PureLibm.CR_ATANF_CN — Constant

Numerator coeffs for rminimax.

polynomials generated using rminimax (https://gitlab.inria.fr/sfilip/rminimax) with the following command:

./ratapprox --function="atan(x)" --dom=[0.000122070,1] --num=[x,x^3,x^5,x^7,x^9,x^11,x^13] --den=[1,x^2,x^4,x^6,x^8,x^10,x^12] --output=atanf.sollya --log

(see output atanf.sollya)

The coefficient cd[0] was slightly reduced from the original value 0x1.51eccde075d67p-2 to avoid an exceptional case for |x| = 0x1.1ad646p-4 and rounding to nearest.

See also CR_ATANF_CD

PureLibm.CR_EXP2F_TB — Constant

[ reinterpret(UInt64, f) for f in CR_EXP2F_TB_F64 ] |> println

PureLibm.CR_TANHF_CD — Constant

Denominator coeffs for rminimax.

See CR_TANHF_CN

PureLibm.CR_TANHF_CN — Constant

Numerator coeffs for rminimax.

polynomials generated using rminimax (https://gitlab.inria.fr/sfilip/rminimax) with the following command:

./ratapprox --function="tanh(x)" --dom=[0x1p-12,9.?????] --num=[x,x^3,x^5,x^7,x^9,x^11,x^13,x^15] --den=[1,x^2,x^4,x^6,x^8,x^10,x^12,x^14] --log --output=tanh.sollya

TODO: find out –dom range

See also CR_TANHF_CD

PureLibm.CR_TGAMMAF_C — Constant

Ploy approximation coefficients for sinpi?

PureLibm.CR_TGAMMAF_TB — Constant

List of exceptional cases.

PureLibm.F32_MAX_NAN — Constant

Name        Hex         hex: %a
NaN (+max)  0x7fffffff  nan

PureLibm.F32_MAX_NORMAL — Constant

Name        Hex         hex: %a             base2               base10(not exact)
Max Normal  0x7f7fffff  0x1.fffffep+127     (1-2^-24)×2^128     3.403E+38

PureLibm.F32_MAX_SUBNORMAL — Constant

Name            Hex         hex: %a         base2               base10(not exact)
Max Subnormal   0x007fffff  0x1.fffffcp-127 (1-2^-23)×2^-126    1.175E-38

PureLibm.F32_MIN_NAN — Constant

Name        Hex         hex: %a
NaN (+min)  0x7f800001  nan

PureLibm.F32_MIN_NORMAL — Constant

Name        Hex         hex: %a     base2   base10(not exact)
Min Normal  0x00800000  0x1p-126    2^-126  1.175E-38

PureLibm.F32_MIN_SUBNORMAL — Constant

Name            Hex         hex: %a     base2   base10(not exact)
Min Subnormal   0x00000001  0x1p-149    2^-149  1.401E-45

PureLibm.F32_POS_INF — Constant

Name        Hex         hex: %a
Inf         0x7f800000  inf

PureLibm.F32_POS_ZERO — Constant

Name    Hex         hex: %a     base2   base10(not exact)
+0      0x00000000  0x0p+0      0       0.000E+00

PureLibm._acosf_as_special — Method

Special cases for acosf when |x| >= 1

PureLibm._asinf_as_special — Method

Special cases for acosf when |x| > 1

PureLibm._llvm_clz — Function

_llvm_clz(x)

Counts the number of leading zeros in x.

// llvm
declare i8   @llvm.ctlz.i8  (i8   <src>, i1 <is_zero_poison>)
@llvm.ctlz.*

// gcc
int __builtin_clz   (unsigned int x)
int __builtin_clzl  (unsigned long)
int __builtin_clzll (unsigned long long)

PureLibm._llvm_popcount — Function

_llvm_popcount(x)

Counts the number of bits set in x.

// llvm
declare i8      @llvm.ctpop.i8(i8  <src>)
declare i16     @llvm.ctpop.i16(i16 <src>)
declare i32     @llvm.ctpop.i32(i32 <src>)
declare i64     @llvm.ctpop.i64(i64 <src>)
declare i256    @llvm.ctpop.i256(i256 <src>)
declare <2 x i32> @llvm.ctpop.v2i32(<2 x i32> <src>)

// clang
__builtin_popcount
__builtin_popcountl
__builtin_popcountll
int __builtin_popcountg(type x)

// gcc
int __builtin_popcount (unsigned int x)
int __builtin_popcountl (unsigned long)
int __builtin_popcountll (unsigned long long)
int __builtin_popcountg (...)

PureLibm._llvm_roundeven — Function

_llvm_roundeven(x)

Round x to the nearest integer in floating-point format rounding halfway cases to even.

// llvm
declare float     @llvm.roundeven.f32(float  %Val)
declare double    @llvm.roundeven.f64(double %Val)
declare x86_fp80  @llvm.roundeven.f80(x86_fp80  %Val)
declare fp128     @llvm.roundeven.f128(fp128 %Val)

// gcc
__builtin_roundeven
__builtin_roundevenf
__builtin_roundevenl

PureLibm._tanf_database — Method

Lookup database, check if tu is a hard to round case, if not return r1.

PureLibm._tanf_rbig — Method

argument reduction same as rltl, but for |x| >= 2^28

PureLibm._tanf_rltl — Method

argument reduction for |z| < 2^28, return r such that 2/pi*x = q + r

PureLibm.as_rsqrt_refine — Method

NOTE: Remvoed support for RoundUp (FE_UPWARD) mode.

PureLibm.cr_acos — Method

cr_acos(x)

Compute arc cosine of x.

PureLibm.cr_acosf — Method

Correctly-rounded arc-cosine function for Float32.

Reference

https://gitlab.inria.fr/core-math/core-math/-/blob/f786e13fb0595adee545d7b29931d283f658ba0a/src/binary32/acos/acosf.c

PureLibm.cr_asin — Method

cr_asin(x)

Compute arc-sine of x.

PureLibm.cr_asinf — Method

Correctly-rounded arc-sine function for Float32.

Reference

https://gitlab.inria.fr/core-math/core-math/-/blob/2c08994e3cd967a63c4c1eed729353a1c3b9c798/src/binary32/asin/asinf.c

PureLibm.cr_atan — Method

cr_atan(x)

Compute arc-tangent of x.

PureLibm.cr_atanf — Method

Correctly-rounded arc-tangent of Float32.

Reference

core-math file commit (a8066a5c)

PureLibm.cr_cos — Method

cr_cos(x)

Compute cosine of x.

PureLibm.cr_cosf — Method

Correctly-rounded cosine of Float32.

PureLibm.cr_exp — Method

cr_exp(x)

Compute natural exponential of x.

PureLibm.cr_exp10 — Method

cr_exp10(x)

Compute 10^x of x.

PureLibm.cr_exp10f — Method

Correctly-rounded 10^x function for Float32.

PureLibm.cr_exp2 — Method

cr_exp2(x)

Compute 2^x of x.

PureLibm.cr_exp2f — Method

Correctly-rounded 2^x function for Float32.

PureLibm.cr_expf — Method

Correctly-rounded natural exponential function for Float32.

PureLibm.cr_log — Method

cr_log(x)

Compute natural logarithm of x.

PureLibm.cr_log1p — Method

cr_log1p(x)

Compute biased argument natural logarithm log(1+x) of x.

PureLibm.cr_log1pf — Method

Correctly-rounded biased argument natural logarithm function for Float32.

PureLibm.cr_logf — Method

Correctly-rounded logarithm function for Float32.

PureLibm.cr_rsqrt — Method

cr_rsqrt(x)

Computes the reciprocal square root of x.

PureLibm.cr_rsqrt — Method

Correctly-rounded reciprocal square root of Float64

PureLibm.cr_rsqrtf — Method

Correctly-rounded reciprocal square root of Float32

PureLibm.cr_sin — Method

cr_sin(x)

Compute sine of x.

PureLibm.cr_sincos — Method

cr_sincos(x)

Compute sine and cosine of x.

PureLibm.cr_sincosf — Method

Correctly-rounded sine and cosine of Float32.

PureLibm.cr_sinf — Method

Correctly-rounded sine of Float32.

PureLibm.cr_tan — Method

cr_tan(x)

Compute tangent of x.

PureLibm.cr_tanf — Method

Correctly-rounded tangent of Float32.

PureLibm.cr_tanh — Method

cr_tanh(x)

Compute hyperbolic tangent of x.

PureLibm.cr_tanhf — Method

Correctly-rounded hyperbolic tangent function for Float32.

PureLibm.cr_tgamma — Method

cr_tgamma(x)

Computes the true gamma function of x.

PureLibm.cr_tgammaf — Method

cr_tgammaf(x::Float32)

Correctly-rounded true gamma function for Float32.

PureLibm.poly12 — Method

Polynomial evaluation for 12 coefficients

PureLibm.poly_x2_c16 — Method

Polynomial evaluation with x^{2n} base and 16 coefficients

PureLibm.@likely — Macro

Hints that the condition cond is likely to be true.

// Macro define in C
#define likely(x)   __builtin_expect(!!(x), 1)

PureLibm.@unlikely — Macro

Hints that the condition cond is likely to be false.

// Macro define in C
#define unlikely(x) __builtin_expect(!!(x), 0)