Function Docs

This list is generated automatically, which may contain some internal unexported functions.

PureLibm.CR_ATANF_CN — Constant

polynomials generated using rminimax (https://gitlab.inria.fr/sfilip/rminimax) with the following command:

./ratapprox --function="atan(x)" --dom=[0.000122070,1] --num=[x,x^3,x^5,x^7,x^9,x^11,x^13] --den=[1,x^2,x^4,x^6,x^8,x^10,x^12] --output=atanf.sollya --log

(see output atanf.sollya)

The coefficient cd[0] was slightly reduced from the original value 0x1.51eccde075d67p-2 to avoid an exceptional case for |x| = 0x1.1ad646p-4 and rounding to nearest.

PureLibm.F32_MAX_NAN — Constant

Name        Hex         hex: %a
NaN (+max)  0x7fffffff  nan

PureLibm.F32_MAX_NORMAL — Constant

Name        Hex         hex: %a             base2               base10(not exact)
Max Normal  0x7f7fffff  0x1.fffffep+127     (1-2^-24)×2^128     3.403E+38

PureLibm.F32_MAX_SUBNORMAL — Constant

Name            Hex         hex: %a         base2               base10(not exact)
Max Subnormal   0x007fffff  0x1.fffffcp-127 (1-2^-23)×2^-126    1.175E-38

PureLibm.F32_MIN_NAN — Constant

Name        Hex         hex: %a
NaN (+min)  0x7f800001  nan

PureLibm.F32_MIN_NORMAL — Constant

Name        Hex         hex: %a     base2   base10(not exact)
Min Normal  0x00800000  0x1p-126    2^-126  1.175E-38

PureLibm.F32_MIN_SUBNORMAL — Constant

Name            Hex         hex: %a     base2   base10(not exact)
Min Subnormal   0x00000001  0x1p-149    2^-149  1.401E-45

PureLibm.F32_POS_INF — Constant

Name        Hex         hex: %a
Inf         0x7f800000  inf

PureLibm.F32_POS_ZERO — Constant

Name    Hex         hex: %a     base2   base10(not exact)
+0      0x00000000  0x0p+0      0       0.000E+00

PureLibm._CR_TGAMMAF_TB — Constant

List of exceptional cases.

PureLibm._llvm_clz — Function

_llvm_clz(x)

Counts the number of leading zeros in x.

// llvm
declare i8   @llvm.ctlz.i8  (i8   <src>, i1 <is_zero_poison>)
@llvm.ctlz.*

// gcc
int __builtin_clz   (unsigned int x)
int __builtin_clzl  (unsigned long)
int __builtin_clzll (unsigned long long)

PureLibm._llvm_roundeven — Function

_llvm_roundeven(x)

Round x to the nearest integer in floating-point format rounding halfway cases to even.

// llvm
declare float     @llvm.roundeven.f32(float  %Val)
declare double    @llvm.roundeven.f64(double %Val)
declare x86_fp80  @llvm.roundeven.f80(x86_fp80  %Val)
declare fp128     @llvm.roundeven.f128(fp128 %Val)

// gcc
__builtin_roundeven
__builtin_roundevenf
__builtin_roundevenl

PureLibm.cr_acos — Method

cr_acos(x)

Compute arc cosine of x.

PureLibm.cr_asin — Method

cr_asin(x)

Compute arc-sine of x.

PureLibm.cr_atan — Method

cr_atan(x)

Compute arc-tangent of x.

PureLibm.cr_atan2 — Method

cr_atan2(y, x)

Compute arc-tangent of y / x.

PureLibm.cr_rsqrt — Method

cr_rsqrt(x)

Computes the reciprocal square root of x.

PureLibm.cr_rsqrt — Method

Correctly-rounded reciprocal square root of Float64

PureLibm.cr_rsqrtf — Method

Correctly-rounded reciprocal square root of Float32

PureLibm.cr_tanh — Method

cr_tanh(x)

Compute hyperbolic tangent of x.

PureLibm.cr_tanhf — Method

Correctly-rounded hyperbolic tangent function of Float32.

PureLibm.cr_tgamma — Method

cr_tgamma(x)

Computes the true gamma function of x.

PureLibm.@likely — Macro

Hints that the condition cond is likely to be true.

// Macro define in C
#define likely(x)   __builtin_expect(!!(x), 1)

PureLibm.@unlikely — Macro

Hints that the condition cond is likely to be false.

// Macro define in C
#define unlikely(x) __builtin_expect(!!(x), 0)